alle Probleme behoben

+++ bildungsgaenge.php
@@ -117,7 +117,7 @@
 					<td>
 						<a href='?page=bildungsgaenge&action=view&bg_id=".$satz['bg_id']."'><img src='icons/view.png' width=15></a>
 						<a href='?page=bildungsgaenge&action=edit&bg_id=".$satz['bg_id']."'><img src='icons/bearbeiten.png' width=15></a>
-						<a href='?page=bildungsgaenge&action=delete&bg_id=".$satz['bg_id']."'><img src='icons/loeschen.png' width=15></a>
+						<a href='?page=bildungsgaenge&action=delete&bg_id=".$satz['bg_id']."' onclick= \"return confirm('Dieser Datensatz wird gelöscht!');\"><img src='icons/loeschen.png' width=15></a>
 					</td>
 				</tr>";
 		}
