+++ klassen.php
@@ -146,7 +146,7 @@
 					<td>
 						<a href='?page=klassen&action=view&k_id=".$satz['k_id']."'><img src='icons/view.png' width=15></a>
 						<a href='?page=klassen&action=edit&k_id=".$satz['k_id']."'><img src='icons/bearbeiten.png' width=15></a>
-						<a href='?page=klassen&action=delete&k_id=".$satz['k_id']."'><img src='icons/loeschen.png' width=15></a>
+						<a href='?page=klassen&action=delete&k_id=".$satz['k_id']."' onclick =\"return confirm('Dieser Datensatz wird gelöscht!');\"><img src='icons/loeschen.png' width=15></a>
 					</td>
 
 				</tr>";
